+++ packages/wix-ui-core/src/testkit/protractor.ts
@@ -6,7 +6,7 @@ import {
 import {
   avatarDriverFactory,
   AvatarDriver,
-} from '../components/avatar/avatar.driver';
+} from '../components/avatar/avatar.uni.driver';
 export const avatarTestkitFactory = protractorUniTestkitFactoryCreator(
   avatarDriverFactory,
 );
